+++ .gitignore
@@ -37,3 +37,6 @@ next-env.d.ts
 
 node_modules
 .env
+
+#home
+src/pages/home/index.js
