+++ src/components/Button.jsx
@@ -1,11 +1,13 @@
 import React from 'react'
 
-const Button = () => {
+import styles from "../styles/Button.module.css"
+
+const Button = ({initBtnSubmit}) => {
   return (
     <>
-      
+      <button className={styles["btn-submit"]}>{initBtnSubmit}</button>
     </>
-  )
+  );
 }
 
 export default Button
