+++ src/pages/auth/forgot.js
@@ -3,10 +3,10 @@ import Image from "next/image";
 
 import styles from "../../styles/Forgot.module.css";
 import logo from "../../assets/Images/tickitz.png";
-import googleIcon from "../../assets/Icons/google.png";
-import facebook from "../../assets/Icons/facebook.png";
+import Button from "../../components/Button";
 
-const forgot = () => {
+
+const Forgot = () => {
   return (
     <>
       <main className={styles["main"]}>
@@ -66,7 +66,7 @@ const forgot = () => {
               </label>
               <input type="checkbox" className={styles["checkbox"]} />
             </span> */}
-            <button className={styles["btn-submit"]}>Join for free Now</button>
+            <Button initBtnSubmit={`Join for free Now`}/>
           </form>
           {/* <p className={styles["direct-to-login"]}>
             Do you already have an account ?{" "}
@@ -96,4 +96,4 @@ const forgot = () => {
   );
 };
 
-export default forgot;
+export default Forgot;
