+++ src/styles/Button.module.css
@@ -1,0 +1,9 @@
+.btn-submit {
+  background: var(--primary-color);
+  color: var(--secondary-color);
+  padding: 1rem 0;
+  border-radius: 14px;
+  font-size: 14px;
+  font-weight: 600;
+  border: none
+}
