+++ src/styles/Forgot.module.css
@@ -3,6 +3,7 @@
   display: flex;
   flex-direction: row;
   min-height: 100vh;
+  font-family: "Inter", sans-serif;
 }
 .left-side {
   /* border: 1px solid darkgreen; */
@@ -138,16 +139,6 @@
   accent-color: var(--primary-color);
   cursor: pointer;
   color: #696f79;
-}
-
-.btn-submit {
-  background: var(--primary-color);
-  color: var(--secondary-color);
-  padding: 1rem 0;
-  border-radius: 14px;
-  font-size: 14px;
-  font-weight: 600;
-  border: none
 }
 
 .direct-to-login {
