+++ src/styles/Login.module.css
@@ -3,6 +3,7 @@
   display: flex;
   flex-direction: row;
   min-height: 100vh;
+  font-family: "Inter", sans-serif;
 }
 .left-side {
   /* border: 1px solid darkgreen; */
@@ -101,16 +102,6 @@
   cursor: pointer;
 }
 
-.btn-submit {
-  background: var(--primary-color);
-  color: var(--secondary-color);
-  padding: 1rem 0;
-  border-radius: 14px;
-  font-size: 14px;
-  font-weight: 600;
-  border: none
-}
-
 .direct-to-reset {
   font-weight: 600;
   color: #696f79;
@@ -124,7 +115,7 @@
 
 .or {
   text-align: center;
-  color: #AAAAAA;
+  color: #aaaaaa;
   font-size: 12px;
 }
