+++ src/styles/globals.css
@@ -1,14 +1,6 @@
 /* || Inter font */
 
-@import url("https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&display=swap");
-/* 
-  {
-  padding: 0;
-  border: 0;
-  margin: 0;
-  box-sizing: border-box;
-  font-family: "Inter", sans-serif;
-} */
+@import url("https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&display=swap"); /* ⇦ Font style global dapat dipakai dihalaman mana saja*/
 
 /* || Root colors*/
 
@@ -16,4 +8,4 @@
   --primary-color: #5f2eea;
   --bg-color: #e5e5e5;
   --secondary-color: #ffffff;
-}
+} /* ⇦ Color style global dapat dipakai dihalaman mana saja*/
